refactor(cron): rename apply_git_changes to materialize_git_mirror
